Backend/messagerie-securisee/requirements.txt
flask
mysql-connector-python
bcrypt
cryptography
